--- backendDTO/src/main/java/ru/vafeen/universityschdeule/backenddto/GroupDTO.kt
@@ -1,7 +1,0 @@
-package ru.vafeen.universityschdeule.backenddto
-
-data class GroupDTO(
-    val id: Int,
-    val course: Int,
-    val group: String
-)
