+++ backendDTO/src/main/java/ru/vafeen/universityschedule/backenddto/GroupDTO.kt
@@ -1,14 +1,14 @@
-package ru.vafeen.universityschedule.domain.models
+package ru.vafeen.universityschedule.backenddto
 
 /**
- * Модель для представления группы в университете.
+ * ДТО (Data Transfer Object) для представления группы в университете.
  * Содержит основную информацию о группе.
  *
  * @property id Уникальный идентификатор группы.
  * @property course Курс, на котором обучается группа.
  * @property group Название группы
  */
-data class Group(
+data class GroupDTO(
     val id: Int,
     val course: Int,
     val group: String
