+++ data/src/main/java/ru/vafeen/universityschedule/data/converters/GroupEntityConverter.kt
@@ -4,13 +4,17 @@ import ru.vafeen.universityschedule.data.database.entity.GroupEntity
 import ru.vafeen.universityschedule.domain.converter.TwoWayBaseConverter
 import ru.vafeen.universityschedule.domain.models.Group
 
+/**
+ * Двусторонний конвертер для преобразования между объектами [GroupEntity] и [Group].
+ * Используется для конвертации данных между форматом сущности базы данных и моделью приложения.
+ */
 internal class GroupEntityConverter : TwoWayBaseConverter<GroupEntity, Group> {
 
     /**
      * Конвертирует объект [GroupEntity] в [Group].
      *
      * @param a Объект [GroupEntity] для конвертации.
-     * @return Объект [Group].
+     * @return Объект [Group], созданный на основе данных из [GroupEntity].
      */
     override fun convertAB(a: GroupEntity): Group {
         return Group(
@@ -24,7 +28,7 @@ internal class GroupEntityConverter : TwoWayBaseConverter<GroupEntity, Group> {
      * Конвертирует объект [Group] в [GroupEntity].
      *
      * @param b Объект [Group] для конвертации.
-     * @return Объект [GroupEntity].
+     * @return Объект [GroupEntity], созданный на основе данных из [Group].
      */
     override fun convertBA(b: Group): GroupEntity {
         return GroupEntity(
