+++ data/src/main/java/ru/vafeen/universityschedule/data/converters/JsonStringTemplateConverter.kt
@@ -10,9 +10,18 @@ import ru.vafeen.universityschedule.domain.models.model_additions.Frequency
 import java.lang.reflect.Type
 import java.time.LocalTime
 
+/**
+ * Класс для конвертации данных между JSON-строками и объектами Kotlin.
+ * Использует Gson для сериализации и десериализации.
+ *
+ * @property frequencyOneWayConverter Конвертер для преобразования частоты занятий из строки.
+ */
 internal class JsonStringTemplateConverter(
     private val frequencyOneWayConverter: FrequencyOneWayConverter
 ) {
+    /**
+     * Экземпляр Gson, настроенный для корректной десериализации LocalTime и Frequency.
+     */
     private var gson: Gson
 
     init {
@@ -21,6 +30,14 @@ internal class JsonStringTemplateConverter(
         // Регистрация десериалайзера для LocalTime
         gsonBuilder.registerTypeAdapter(LocalTime::class.java, object :
             JsonDeserializer<LocalTime> {
+            /**
+             * Десериализует строку в объект LocalTime.
+             *
+             * @param json JSON-элемент для десериализации.
+             * @param typeOfT Тип объекта, который нужно получить.
+             * @param context Контекст десериализации.
+             * @return Объект LocalTime, полученный из строки.
+             */
             override fun deserialize(
                 json: JsonElement?,
                 typeOfT: Type?,
@@ -31,6 +48,15 @@ internal class JsonStringTemplateConverter(
         // Регистрация десериалайзера для Frequency
         gsonBuilder.registerTypeAdapter(Frequency::class.java, object :
             JsonDeserializer<Frequency> {
+            /**
+             * Десериализует строку в объект Frequency.
+             * Если строка пустая, возвращает Frequency.Every по умолчанию.
+             *
+             * @param json JSON-элемент для десериализации.
+             * @param typeOfT Тип объекта, который нужно получить.
+             * @param context Контекст десериализации.
+             * @return Объект Frequency, полученный из строки.
+             */
             override fun deserialize(
                 json: JsonElement?,
                 typeOfT: Type?,
@@ -42,10 +68,21 @@ internal class JsonStringTemplateConverter(
         gson = gsonBuilder.create()
     }
 
+    /**
+     * Конвертирует JSON-строку в объект типа T.
+     *
+     * @param jsonString JSON-строка для конвертации.
+     * @return Объект типа T.
+     */
     inline fun <reified T> convertObject(jsonString: String): T =
         gson.fromJson(jsonString, T::class.java)
 
+    /**
+     * Конвертирует JSON-строку в список объектов типа T.
+     *
+     * @param jsonString JSON-строка для конвертации.
+     * @return Список объектов типа T.
+     */
     inline fun <reified T> convertList(jsonString: String): List<T> =
         gson.fromJson(jsonString, object : TypeToken<List<T>>() {}.type)
-
 }
