+++ data/src/main/java/ru/vafeen/universityschedule/data/impl/network/repository/LessonDataRepositoryImpl.kt
@@ -7,16 +7,22 @@ import ru.vafeen.universityschedule.domain.models.Lesson
 import ru.vafeen.universityschedule.domain.network.repository.LessonDataRepository
 import ru.vafeen.universityschedule.domain.network.result.ResponseResult
 
+/**
+ * Реализация репозитория для работы с данными занятий через сеть.
+ * Использует сервис [LessonDataService] для запросов к серверу и конвертер [LessonDTOConverter] для преобразования данных.
+ *
+ * @property lessonDTOConverter Конвертер для преобразования данных из формата DTO в модель приложения.
+ * @property lessonDataService Сервис для запросов к серверу.
+ */
 internal class LessonDataRepositoryImpl(
     private val lessonDTOConverter: LessonDTOConverter,
     private val lessonDataService: LessonDataService,
 ) : LessonDataRepository {
 
-
     /**
      * Получает список занятий для указанной группы с сервера.
      *
-     * @param group Идентификатор группы, для которой требуется получить расписание.
+     * @param groupId Идентификатор группы, для которой требуется получить расписание.
      * @return [ResponseResult] с результатом запроса, содержащим список занятий ([Lesson]) или информацию об ошибке.
      */
     override suspend fun getLessonDataByGroupId(groupId: Int): ResponseResult<List<Lesson>> =
@@ -30,9 +36,13 @@ internal class LessonDataRepositoryImpl(
             )
         }
 
+    /**
+     * Получает полный список всех занятий с сервера.
+     *
+     * @return [ResponseResult] с результатом запроса, содержащим список занятий ([Lesson]) или информацию об ошибке.
+     */
     override suspend fun getAllLessonData(): ResponseResult<List<Lesson>> =
         getResponseWrappedAllErrors {
             ResponseResult.Success(lessonDTOConverter.convertList(lessonDataService.getLessonData()))
         }
-
 }
