+++ data/src/main/java/ru/vafeen/universityschedule/data/network/service/server/LessonDataService.kt
@@ -3,7 +3,7 @@ package ru.vafeen.universityschedule.data.network.service.server
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
-import ru.vafeen.universityschdeule.backenddto.LessonDTO
+import ru.vafeen.universityschedule.backenddto.LessonDTO
 import ru.vafeen.universityschedule.data.converters.JsonStringTemplateConverter
