+++ domain/src/main/java/ru/vafeen/universityschedule/domain/models/Lesson.kt
@@ -5,7 +5,8 @@ import java.time.DayOfWeek
 import java.time.LocalTime
 
 /**
- * Класс, представляющий пару.
+ * Модель для представления пары в расписании.
+ * Содержит подробную информацию о паре, включая день недели, время начала и конца, аудиторию, преподавателя и другие атрибуты.
  *
  * @property id Уникальный идентификатор пары.
  * @property dayOfWeek День недели, в который проходит пара (может быть null).
@@ -19,6 +20,7 @@ import java.time.LocalTime
  * @property idOfReminderBeforeLesson Идентификатор напоминания перед парой (может быть null).
  * @property idOfReminderAfterBeginningLesson Идентификатор напоминания после начала пары (может быть null).
  * @property note Примечание к паре (может быть null).
+ * @property linkToCourse Ссылка на курс, связанный с парой (может быть null).
  */
 data class Lesson(
     val id: Int = 0,
@@ -38,9 +40,8 @@ data class Lesson(
 
     /**
      * Возвращает строковое представление объекта [Lesson].
-     *
-     * @return Строка с информацией о дне недели, названии, времени начала и окончания,
-     * аудитории, преподавателе, подгруппе и частоте.
+     * Включает информацию о дне недели, названии, времени начала и окончания,
+     * аудитории, преподавателе, подгруппе, частоте и ссылке на курс.
      */
     override fun toString(): String {
         return "\n dayOfWeek=${dayOfWeek ?: "\"is null\""} name=${name ?: "\"is null\""} st=${startTime}-et=${endTime} classroom=${classroom ?: "\"is null\""} tchr=${teacher ?: "\"is null\""} sbgr=${subGroup ?: "\"is null\""} fr=${frequency ?: "\"is null\""} linkToCourse=${linkToCourse ?: "\"is null\""}"
@@ -50,7 +51,7 @@ data class Lesson(
      * Сравнивает текущий объект [Lesson] с другим объектом [Lesson] по времени начала.
      *
      * @param other Другой объект [Lesson] для сравнения.
-     * @return 1, если время начала больше; 0, если равно; -1 в противном случае.
+     * @return 1, если время начала текущей пары позже; 0, если равно; -1, если раньше.
      */
     override fun compareTo(other: Lesson): Int = when {
         startTime > other.startTime -> 1
