+++ domain/src/main/java/ru/vafeen/universityschedule/domain/models/model_additions/Frequency.kt
@@ -2,7 +2,12 @@ package ru.vafeen.universityschedule.domain.models.model_additions
 
 import ru.vafeen.universityschedule.resources.R
 
-
+/**
+ * Перечисление для представления частоты проведения занятий.
+ * Каждый вариант имеет ресурс строки для отображения в интерфейсе.
+ *
+ * @property resourceName Идентификатор ресурса строки для отображения частоты.
+ */
 enum class Frequency(val resourceName: Int) {
     Every(resourceName = R.string.every) {
         override fun getOpposite(): Frequency = Every
@@ -14,5 +19,9 @@ enum class Frequency(val resourceName: Int) {
         override fun getOpposite(): Frequency = Numerator
     };
 
+    /**
+     * Абстрактный метод для получения противоположной частоты.
+     * Должен быть реализован для каждого варианта перечисления.
+     */
     abstract fun getOpposite(): Frequency
 }
