--- domain/src/main/java/ru/vafeen/universityschedule/domain/network/GoogleSheetsServiceLink.kt
@@ -1,5 +1,0 @@
-package ru.vafeen.universityschedule.domain.network
-
-object GoogleSheetsServiceLink {
-    const val BASE_URL = "https://docs.google.com/"
-}
